move db layer from lltk/tools/ to lltk/db/
14 files renamed with cleaner names:
  tools/db_adapter.py         → db/adapter.py
  tools/metadb.py             → db/metadb.py
  tools/metadb_ch.py          → db/metadb_ch.py
  tools/clickhouse_schema.py  → db/schema.py
  tools/clickhouse_rebuild.py → db/rebuild.py
  tools/clickhouse_ingest.py  → db/ingest.py
  tools/clickhouse_match.py   → db/match.py
  tools/clickhouse_enrich.py  → db/enrich.py
  tools/clickhouse_detect_langs.py → db/detect_langs.py
  tools/clickhouse_text_words.py   → db/text_words.py
  tools/clickhouse_wordindex.py    → db/wordindex.py
  tools/clickhouse_passages.py     → db/passages.py
  tools/clickhouse_embeddings.py   → db/embeddings.py
  tools/db_migrate.py         → db/migrate.py

All import paths updated across 17 source files and 7 test files.
lltk.db is now the single home for the ClickHouse database layer.
