+++ lltk/db/__init__.py
@@ -1,0 +1,1 @@
+from .metadb import metadb as db
