+++ lltk/db/passages.py
@@ -115,7 +115,7 @@ def build_passages_ch(adapter, n: int = 500, num_proc: int | None = None,
     Used by MetaDBCH.build_passages_db when year/dedup/SyntheticCorpus filtering
     has already resolved the exact set of texts to chunk.
     """
-    from lltk.tools.metadb import _chunk_text_to_passages, DB_BLACKLIST
+    from lltk.db.metadb import _chunk_text_to_passages, DB_BLACKLIST
     from lltk.tools.tools import pmap, get_tqdm
     from lltk.tools.vocabs import normalize_lang
