+++ lltk/db/wordindex.py
